ubuntu/API/requirements.txt
flask
python-magic
numpy
pydub
librosa
python-multipart
ffmpeg
werkzeug
soundfile
pandas
matplotlib
sklearn
statistics
